zcash_client_sqlite: Factor out common note selection code.
